% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode.R
\name{barcode_print}
\alias{barcode_print}
\title{Print barcodes}
\usage{
barcode_print(code)
}
\arguments{
\item{code}{Value to print as barcode}
}
\description{
Use zint API to print barcodes
}
\examples{
\dontrun{
zintr::barcode_print("Hello World")
}

}
